refactor(menuProductRequestFixture): menuProductRequestFixture를 class화하라
